fs/binfmt_elf: fix memory map for PIE applications
arch/x86/include/asm/elf.h comment says:
"
ELF_ET_DYN_BASE is the location that an ET_DYN program is loaded
if exec'ed.  Typical use of this is to invoke "./ld.so someprog"
to test out a new version of the loader.  We need to make sure
that it is out of the way of the program that it will "exec",
and that there is sufficient room for the brk.
"

In case we have main application linked as PIE, this will cause
problems as the main program itself will go the memory location
that allows very little heap.

This fixes the loader to detect PIE application by checking if
elf_interpreter is requested, and loads them to beginning of the
address space instead of the specially crafted place for elf
interpreter.

Signed-off-by: Timo Teräs <timo.teras@iki.fi>